feat: split up capability structure (#11)
This changes the capability definition to support all existing features, and be extensible.

The client constructor for now just creates a token with an All capability. We'll need to add a capability builder likely, but that doesn't have to be in the first round of changes (and in a way it only makes sense once we add delegation).

Note that this is a breaking change for the RPC


* feat: split up capability structure

* deps: use patched rcan

* feat: add some ergonomics improvements

* chore: use rcan main again

* refactor: flatten caps structure, add to-and-from string conversions

* feat: add GetBlob cap

* fix: cleanups