feat!(tickets): tickets API (#54)
Add an API for sharing tickets via n0des.

This started off as signals, and then after a bunch of feedback we've evolved it to center around tickets. Tickets work well because we can leverage the existing trait a bunch, and use it to encourage users not to rely solely on n0des for ticket passing.

This API supports publishing multiple types of tickets, and uses a generic to handle decoding the wire response for the user, which I think makes for very nice ergonomics.

The choice to rely on metrics interval sending as a keepalive for published tickets is admittedly strange, but meh, it's here & it works.

* feat(signals): initial cut at a signaling API
* sketch out an example
* feat(tickets): API for working with tickets
* feat(caps): basic ticket capabilities
* chore(deps): bump past yanked crates
* fix(protocol): wrap Ticket irpc method responses in RemoteResult
* auth on ticket send/receive
* document ticket publication
* tickets are disabled by default
* fix feature gating
* remove ssh example
* unpublish tickets as well
* example cleanup
* feat: fetch ticket by name
* add fetch_ticket to tickets example
* show ticket timeout in example